Create clash HTML (index.html)
if you would like me to add comments to make it easier to differentiate between divs please let me know and i can reload with comments. I think also if we will be uploading to GitHub updated code we should notify everyone because its one file and we dont want to be working on something in an outdated version of the file.